Added scene flow and a more complex UI
Very big commit:
Now we can die and revive then
or we can pass to next level
or we can get an game over and reset the game.

Also we have: Main menu, and a more complex UI that helps us with:
Points, Lives, and the level in which we are.

And UI also show you messages like: LEVEL X, READY, GO!
LEVEL COMPLETE!
GAME OVER!